+++ src/lib/components/data-table/DataTableEvent.svelte
@@ -40,8 +40,12 @@
 	};
 
 	const traductions: { [key: string]: string } = {
-		ringside_tickets: 'Ringside',
-		general_tickets: 'General'
+		success: 'Web',
+		system: 'Sistema',
+		register: 'Registro',
+		rejected: 'Rechazado',
+		refund: 'Reembolso',
+		change: 'Cambio'
 	};
 
 	export let Payments: Payment[];
@@ -56,7 +60,7 @@
 		});
 	}
 
-	let showOnlySuccess = writable(false);
+	let showRejected = writable(false);
 
 	// Reactive statement to filter payments based on search term and switch state
 	$: filteredPayments = Payments.filter((payment) => {
@@ -67,20 +71,43 @@
 			payment.payment_status?.toLowerCase().includes(searchTerm.toLowerCase()) ||
 			payment.rut?.toLowerCase().includes(searchTerm.toLowerCase());
 
-		const matchesSuccessFilter = $showOnlySuccess || payment.payment_status === 'success';
+		const matchesSuccessFilter = $showRejected || payment.payment_status === 'success' || payment.payment_status === 'system' || payment.payment_status === 'refund' || payment.payment_status === 'change';
+
+		if ($showRejected) {
+			return (payment.payment_status === 'register' || payment.payment_status === 'rejected' || payment.payment_status === null) && matchesSearchTerm;
+		}
 
 		return matchesSearchTerm && matchesSuccessFilter;
 	});
+
+	// Function to determine the badge class based on payment status
+	function getBadgeClass(paymentStatus: string): string {
+		switch (paymentStatus) {
+			case 'success':
+				return 'bg-orange-300 text-orange-900';
+			case 'rejected':
+				return 'bg-text-300 bg-text-800';
+			case 'system':
+				return 'bg-blue-300 text-blue-900';
+			case 'register':
+				return 'bg-gray-300 text-gray-900';
+			case 'refund':
+				return 'bg-red-300 text-red-900';
+			case 'change':
+				return 'bg-yellow-300 text-yellow-900';
+			default:
+				return 'bg-gray-400';
+		}
+	}
 </script>
 
 <!-- HEADER -->
 <div class="flex flex-col md:flex-row items-center py-4 md:gap-4 justify-between">
 	<div class="flex flex-col md:flex-row md:items-center gap-4 items-left w-full">
-		<h2 class="font-bold text-xl">Listado de compras</h2>
-		<Input type="text" placeholder="Buscador..." class="w-full md:w-96" bind:value={searchTerm} />
+		<Input type="text" placeholder="Buscador..." class="w-full md:w-96 border-primary" bind:value={searchTerm} />
 		<div class="flex items-center space-x-2">
-			<Switch id="rejected-payments" bind:checked={$showOnlySuccess} />
-			<Label for="rejected-payments">Incluir pagos rechazados</Label>
+			<Switch id="rejected-payments" bind:checked={$showRejected} />
+			<Label for="rejected-payments">Ver pagos rechazados</Label>
 		</div>
 	</div>
 	<DialogToAddPayments />
@@ -93,7 +120,7 @@
 			<li>
 				<button
 					on:click={() => openDialog(payment.id)}
-					class="p-6 w-full hover:bg-muted flex justify-between"
+					class={`p-6 w-full hover:bg-muted flex justify-between ${payment.ticketAmount === payment.ticketValidated ? 'bg-green-500/10' : ''}`}
 				>
 					<div class="flex flex-col items-start">
 						<span class="text-xs text-primary uppercase">
@@ -117,13 +144,9 @@
 					</div>
 					<div class="flex flex-col items-end gap-4">
 						<Badge
-							class={`${
-								payment.payment_status === 'success'
-									? 'bg-yellow-400'
-									: 'bg-red-400 hover:bg-red-500'
-							}`}
+							class={getBadgeClass(payment.payment_status)}
 						>
-							{payment.payment_status}
+							{traductions[payment.payment_status]}
 						</Badge>
 						{#if payment.ticketAmount === payment.ticketValidated}
 							<Badge class="bg-green-400 hover:bg-green-500">validado</Badge>