+++ src/routes/+layout.server.ts
@@ -5,6 +5,7 @@ import type { PageServerLoad } from './$types';
 export const load: PageServerLoad = async ({ locals }) => {
 	if (!locals.session) throw redirect(302, '/login');
 	return {
-		validator: locals.session.user?.public_metadata.validator
+		validator: locals.session.user?.public_metadata.validator,
+		admin: locals.session.user?.public_metadata.admin
 	};
 };